+++ packages/ui-transfer/src/components/depositForm/swap/useSlippage.ts
@@ -1,0 +1,12 @@
+import { useLocalStorage } from "@veltodefi/hooks";
+
+const ORDERLY_SWAP_DEPOSIT_SLIPPAGE_KEY = "orderly_swap_deposit_slippage";
+
+export const useSlippage = () => {
+  const [slippage, setSlippage] = useLocalStorage(
+    ORDERLY_SWAP_DEPOSIT_SLIPPAGE_KEY,
+    1,
+  );
+
+  return { slippage, onSlippageChange: setSlippage };
+};