+++ packages/storybook-theme-tool/CHANGELOG.md
@@ -1,5 +1,60 @@
 # storybook-theme-tool
 
+## 2.10.0
+
+### Minor Changes
+
+- e8fcd4a: publish
+- 12142ab: publish
+- publish
+
+### Patch Changes
+
+- Updated dependencies [e8fcd4a]
+- Updated dependencies [12142ab]
+- Updated dependencies
+  - @veltodefi/ui@2.10.0
+
+## 2.10.0-alpha.1
+
+### Minor Changes
+
+- publish
+
+### Patch Changes
+
+- Updated dependencies
+  - @veltodefi/ui@2.10.0-alpha.1
+
+## 2.10.0-alpha.0
+
+### Minor Changes
+
+- publish
+
+### Patch Changes
+
+- Updated dependencies
+  - @veltodefi/ui@2.10.0-alpha.0
+
+## 2.9.1
+
+### Patch Changes
+
+- publish
+- f22ab39: publish
+- Updated dependencies
+- Updated dependencies [f22ab39]
+  - @veltodefi/ui@2.9.1
+
+## 2.9.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @veltodefi/ui@2.9.1-alpha.0
+
 ## 2.9.0
 
 ### Minor Changes